[refactor] 리뷰 DTO 디렉토리 변경 (#113)
