+++ src/main/java/org/recordy/server/common/util/data/ExhibitionDataInitializer.java
@@ -33,7 +33,7 @@ import java.util.Objects;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 @Slf4j
-@Profile({"dev", "local"})
+@Profile({"dev"})
 @Component
 public class ExhibitionDataInitializer {
