+++ src/main/java/org/recordy/server/place/service/dto/google/GooglePlaceDetails.java
@@ -1,7 +1,5 @@
 package org.recordy.server.place.service.dto.google;
 
-import org.recordy.server.place.service.dto.Review;
-
 import java.util.List;
 
 public record GooglePlaceDetails(
