+++ src/main/java/org/recordy/server/place/service/dto/google/Review.java
@@ -1,4 +1,4 @@
-package org.recordy.server.place.service.dto;
+package org.recordy.server.place.service.dto.google;
 
 public record Review(
         String author_name,
